split attributes
