+++ src/tools/clippy/clippy_utils/src/msrvs.rs
@@ -1,4 +1,4 @@
-use rustc_attr::AttributeExt;
+use rustc_ast::attr::AttributeExt;
 use rustc_attr::parse_version;
 use rustc_session::{RustcVersion, Session};
 use rustc_span::{Symbol, sym};
